+++ src/index.ts
@@ -11,9 +11,12 @@ import { CheckboxGroup } from './components/form/checkboxes/checkbox-group/check
 import ContentEditableInput from './components/form/inputs/content-editable-input/content-editable-input.component';
 import { DateTime } from './components/form/date-time/date-time.component';
 import FormattedInput from './components/form/inputs/formatted-input/formatted-input.component';
+import { GroupButton } from './components/form/buttons/group-button/group-button.component.tsx';
 import { MediaQuery } from './components/mobile/media-query/media-query.component.tsx';
 import { MobileOverlayPanel } from './components/mobile/overlay/mobile-overlay-panel.component.tsx';
 import OverlayPanel from './components/overlay/overlay-panel/overlay-panel.component';
+import { RadioButton } from './components/form/radio-buttons/radio-button/radio-button.component.tsx';
+import { Toggle } from './components/form/checkboxes/toggle/toggle.component.tsx';
 
 // functions
 import { createBeeSoftTheme, applyBeeSoftTheme } from './components/common-functions';
@@ -25,6 +28,8 @@ import { DateFormatType, DateSelectionType } from './components/form/date-time/d
 import { DateTimeInputTemplateProps } from './components/form/date-time/date-time.props.ts';
 import { DateTimeScrollerTemplateProps } from './components/form/date-time/date-time-scroller.component';
 import { DateTimeCalendarTemplateProps } from './components/form/date-time/date-time-calendar.component';
+import { GroupButtonItemTemplateProps } from './components/form/buttons/group-button/group-button.props.ts';
+import { RadioChangeEvent } from './components/form/radio-buttons/radio-button/radio-button.props.ts';
 
 // styles
 import './index.css';
@@ -37,9 +42,12 @@ export {
   ContentEditableInput,
   DateTime,
   FormattedInput,
+  GroupButton,
   MediaQuery,
   MobileOverlayPanel,
   OverlayPanel,
+  RadioButton,
+  Toggle,
 
   // functions
   applyBeeSoftTheme,
@@ -59,4 +67,6 @@ export type {
   DateTimeInputTemplateProps,
   DateTimeScrollerTemplateProps,
   DateTimeCalendarTemplateProps,
+  GroupButtonItemTemplateProps,
+  RadioChangeEvent,
 };